Fixed updating issue for info sent to db

+++ Cobit-19/Shared/Dtos/ComponentQuestionDto.cs
@@ -5,7 +5,7 @@
         public string questionType { get; set; }
         public string questionDescription { get; set; }
         public int questionAnswer { get; set; }
-        public float questionScore { get; set; }
+        public double questionScore { get; set; }
         public string? questionComment { get; set; }
         public string? questionSkill { get; set; }
         public string? relatedGuidance { get; set; }
